register platform api during initialize
