+++ library/common/extensions/cert_validator/platform_bridge/config.cc
@@ -10,8 +10,10 @@ namespace Tls {
 CertValidatorPtr PlatformBridgeCertValidatorFactory::createCertValidator(
     const Envoy::Ssl::CertificateValidationContextConfig* config, SslStats& stats,
     TimeSource& /*time_source*/) {
-  platform_validator_ =
-      static_cast<envoy_cert_validator*>(Api::External::retrieveApi("platform_cert_validator"));
+  if (platform_validator_ == nullptr) {
+    platform_validator_ =
+        static_cast<envoy_cert_validator*>(Api::External::retrieveApi("platform_cert_validator"));
+  }
   return std::make_unique<PlatformBridgeCertValidator>(config, stats, platform_validator_);
 }
