+++ library/common/jni/BUILD
@@ -210,10 +210,12 @@ cc_library(
     deps = [
         "@envoy//source/common/common:assert_lib",
         "//library/common/types:c_types_lib",
+        ":jni_utility_lib",
+        "//library/common/extensions/cert_validator/platform_bridge:c_types_lib",
+        "//library/common/api:c_types",
     ] + select({
         "@envoy//bazel:android": [
             ":ndk_jni_support",
-            ":jni_utility_lib",
             "//library/common/data:utility_lib",
             "//library/common/jni/import:jni_import_lib",
         ],
