+++ library/common/jni/android_test_jni_interface.cc
@@ -1,3 +1,4 @@
+#include "library/common/jni/android_jni_utility.h"
 #include "library/common/jni/import/jni_import.h"
 #include "library/common/jni/jni_support.h"
 #include "library/common/jni/jni_utility.h"
@@ -13,6 +14,11 @@ Java_io_envoyproxy_envoymobile_engine_AndroidJniLibrary_initialize(JNIEnv* env,
                                                                    jobject class_loader,
                                                                    jobject connectivity_manager) {
   set_class_loader(env->NewGlobalRef(class_loader));
-
+  // At this point, we know Android APIs are available. Register cert chain validation JNI calls.
+  envoy_status_t result =
+      register_platform_api(cert_validator_name, get_android_cert_validator_api());
+  if (result == ENVOY_FAILURE) {
+    return -1;
+  }
   return 0;
 }
