+++ library/common/jni/jni_interface.cc
@@ -1,22 +1,27 @@
 #include <ares.h>
 
-#include <iostream>
-#include <string>
-#include <vector>
-
 #include "library/common/api/c_types.h"
 #include "library/common/data/utility.h"
-#include "library/common/extensions/cert_validator/platform_bridge/c_types.h"
 #include "library/common/extensions/filters/http/platform_bridge/c_types.h"
 #include "library/common/extensions/key_value/platform/c_types.h"
+#include "library/common/jni/android_jni_utility.h"
 #include "library/common/jni/import/jni_import.h"
 #include "library/common/jni/jni_support.h"
 #include "library/common/jni/jni_utility.h"
 #include "library/common/jni/jni_version.h"
 #include "library/common/main_interface.h"
-#include "openssl/ssl.h"
 
 // NOLINT(namespace-envoy)
+
+JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM* vm, void* reserved) {
+  JNIEnv* env = nullptr;
+  if (vm->GetEnv(reinterpret_cast<void**>(&env), JNI_VERSION) != JNI_OK) {
+    return -1;
+  }
+
+  set_vm(vm);
+  return JNI_VERSION;
+}
 
 // JniLibrary
 
@@ -1153,155 +1158,6 @@ extern "C" JNIEXPORT jint JNICALL Java_io_envoyproxy_envoymobile_engine_JniLibra
   return result;
 }
 
-// EnvoyCertValidator
-bool jvm_cert_is_issued_by_known_root(JNIEnv* env, jobject result) {
-  jclass jcls_AndroidCertVerifyResult = find_class("org.chromium.net.AndroidCertVerifyResult");
-  jmethodID jmid_isIssuedByKnownRoot =
-      env->GetMethodID(jcls_AndroidCertVerifyResult, "isIssuedByKnownRoot", "()Z");
-  ASSERT(jmid_isIssuedByKnownRoot);
-  bool is_issued_by_known_root = env->CallBooleanMethod(result, jmid_isIssuedByKnownRoot);
-  env->DeleteLocalRef(jcls_AndroidCertVerifyResult);
-  return is_issued_by_known_root;
-}
-
-envoy_cert_verify_status_t jvm_cert_get_status(JNIEnv* env, jobject j_result) {
-  jclass jcls_AndroidCertVerifyResult = find_class("org.chromium.net.AndroidCertVerifyResult");
-  jmethodID jmid_getStatus = env->GetMethodID(jcls_AndroidCertVerifyResult, "getStatus", "()I");
-  ASSERT(jmid_getStatus);
-  envoy_cert_verify_status_t result = CERT_VERIFY_STATUS_FAILED;
-  result = static_cast<envoy_cert_verify_status_t>(env->CallIntMethod(j_result, jmid_getStatus));
-
-  env->DeleteLocalRef(jcls_AndroidCertVerifyResult);
-  return result;
-}
-
-jobjectArray jvm_cert_get_certificate_chain_encoded(JNIEnv* env, jobject result) {
-  jclass jcls_AndroidCertVerifyResult = find_class("org.chromium.net.AndroidCertVerifyResult");
-  jmethodID jmid_getCertificateChainEncoded =
-      env->GetMethodID(jcls_AndroidCertVerifyResult, "getCertificateChainEncoded", "()[[B");
-  jobjectArray certificate_chain =
-      static_cast<jobjectArray>(env->CallObjectMethod(result, jmid_getCertificateChainEncoded));
-  env->DeleteLocalRef(jcls_AndroidCertVerifyResult);
-  return certificate_chain;
-}
-
-// Once we have a better picture of how Android's certificate verification will
-// be plugged into EM, we should decide where this function should really live.
-// Context: as of now JNI functions declared in this file are not exported through any
-// header files, instead they are stored as callbacks into plain function
-// tables. For this reason, this function, which would ideally be defined in
-// jni_utility.cc, is currently defined here.
-static void ExtractCertVerifyResult(JNIEnv* env, jobject result, envoy_cert_verify_status_t* status,
-                                    bool* is_issued_by_known_root,
-                                    std::vector<std::string>* verified_chain) {
-  *status = jvm_cert_get_status(env, result);
-  if (*status == CERT_VERIFY_STATUS_OK) {
-    *is_issued_by_known_root = jvm_cert_is_issued_by_known_root(env, result);
-    jobjectArray chain_byte_array = jvm_cert_get_certificate_chain_encoded(env, result);
-    if (chain_byte_array != nullptr) {
-      JavaArrayOfByteArrayToStringVector(env, chain_byte_array, verified_chain);
-    }
-  }
-}
-
-// `auth_type` and `host` are expected to be UTF-8 encoded.
-static jobject call_jvm_verify_x509_cert_chain(JNIEnv* env,
-                                               const std::vector<std::string>& cert_chain,
-                                               std::string auth_type, std::string host) {
-  jni_log("[Envoy]", "jvm_verify_x509_cert_chain");
-  jclass jcls_AndroidNetworkLibrary = find_class("org.chromium.net.AndroidNetworkLibrary");
-  jmethodID jmid_verifyServerCertificates =
-      env->GetStaticMethodID(jcls_AndroidNetworkLibrary, "verifyServerCertificates",
-                             "([[B[B[B)Lorg/chromium/net/AndroidCertVerifyResult;");
-  jobjectArray chain_byte_array = ToJavaArrayOfByteArray(env, cert_chain);
-  jbyteArray auth_string = ToJavaByteArray(env, auth_type);
-  jbyteArray host_string = ToJavaByteArray(env, host);
-  jobject result =
-      env->CallStaticObjectMethod(jcls_AndroidNetworkLibrary, jmid_verifyServerCertificates,
-                                  chain_byte_array, auth_string, host_string);
-  env->DeleteLocalRef(chain_byte_array);
-  env->DeleteLocalRef(auth_string);
-  env->DeleteLocalRef(host_string);
-  env->DeleteLocalRef(jcls_AndroidNetworkLibrary);
-  return result;
-}
-
-// `auth_type` and `host` are expected to be UTF-8 encoded.
-static void jvm_verify_x509_cert_chain(const std::vector<std::string>& cert_chain,
-                                       std::string auth_type, std::string host,
-                                       envoy_cert_verify_status_t* status,
-                                       bool* is_issued_by_known_root,
-                                       std::vector<std::string>* verified_chain) {
-  JNIEnv* env = get_env();
-  jobject result = call_jvm_verify_x509_cert_chain(env, cert_chain, auth_type, host);
-  if (env->ExceptionCheck() == JNI_TRUE) {
-    env->ExceptionDescribe();
-    env->ExceptionClear();
-    *status = CERT_VERIFY_STATUS_NOT_YET_VALID;
-  } else {
-    ExtractCertVerifyResult(get_env(), result, status, is_issued_by_known_root, verified_chain);
-    if (env->ExceptionCheck() == JNI_TRUE) {
-      env->ExceptionDescribe();
-      env->ExceptionClear();
-      *status = CERT_VERIFY_STATUS_FAILED;
-    };
-  }
-  env->DeleteLocalRef(result);
-}
-
-static envoy_cert_validation_result verify_x509_cert_chain(const envoy_data* certs, uint8_t size,
-                                                           const char* host_name) {
-  jni_log("[Envoy]", "verify_x509_cert_chain");
-
-  envoy_cert_verify_status_t result;
-  bool is_issued_by_known_root;
-  std::vector<std::string> verified_chain;
-  std::vector<std::string> cert_chain;
-  for (uint8_t i = 0; i < size; ++i) {
-    cert_chain.push_back(Envoy::Data::Utility::copyToString(certs[i]));
-    release_envoy_data(certs[i]);
-  }
-
-  // Android ignores the authType parameter to X509TrustManager.checkServerTrusted, so pass in "RSA"
-  // as dummy value. See https://crbug.com/627154.
-  jvm_verify_x509_cert_chain(cert_chain, "RSA", host_name, &result, &is_issued_by_known_root,
-                             &verified_chain);
-  switch (result) {
-  case CERT_VERIFY_STATUS_OK:
-    return {ENVOY_SUCCESS};
-  case CERT_VERIFY_STATUS_EXPIRED: {
-    return {ENVOY_FAILURE, SSL_AD_CERTIFICATE_EXPIRED,
-            "AndroidNetworkLibrary_verifyServerCertificates failed: expired cert."};
-  }
-  case CERT_VERIFY_STATUS_NO_TRUSTED_ROOT:
-    return {ENVOY_FAILURE, SSL_AD_CERTIFICATE_UNKNOWN,
-            "AndroidNetworkLibrary_verifyServerCertificates failed: no trusted root."};
-  case CERT_VERIFY_STATUS_UNABLE_TO_PARSE:
-    return {ENVOY_FAILURE, SSL_AD_BAD_CERTIFICATE,
-            "AndroidNetworkLibrary_verifyServerCertificates failed: unable to parse cert."};
-  case CERT_VERIFY_STATUS_INCORRECT_KEY_USAGE:
-    return {ENVOY_FAILURE, SSL_AD_CERTIFICATE_UNKNOWN,
-            "AndroidNetworkLibrary_verifyServerCertificates failed: incorrect key usage."};
-  case CERT_VERIFY_STATUS_FAILED:
-    return {
-        ENVOY_FAILURE, SSL_AD_CERTIFICATE_UNKNOWN,
-        "AndroidNetworkLibrary_verifyServerCertificates failed: validation couldn't be conducted."};
-  case CERT_VERIFY_STATUS_NOT_YET_VALID:
-    return {ENVOY_FAILURE, SSL_AD_CERTIFICATE_UNKNOWN,
-            "AndroidNetworkLibrary_verifyServerCertificates failed: not yet valid."};
-  }
-}
-
-envoy_status_t jvm_register_cert_validator() {
-  jni_log("[Envoy]", "registerCertValidatorFactory");
-  // TODO(danzh) this object leaks, but it's tied to the life of the engine.
-  envoy_cert_validator* api = (envoy_cert_validator*)safe_malloc(sizeof(envoy_cert_validator));
-  api->validate_cert = verify_x509_cert_chain;
-  api->release_validator = jvm_detach_thread;
-  const std::string platform_name = "platform_cert_validator";
-  return register_platform_api(platform_name.c_str(), api);
-}
-
 static void jvm_add_test_root_certificate(const uint8_t* cert, size_t len) {
   jni_log("[Envoy]", "jvm_add_test_root_certificate");
   JNIEnv* env = get_env();
@@ -1352,19 +1208,4 @@ extern "C" JNIEXPORT void JNICALL
 Java_io_envoyproxy_envoymobile_engine_JniLibrary_callClearTestRootCertificateFromNative(JNIEnv*,
                                                                                         jclass) {
   jvm_clear_test_root_certificate();
-}
-
-JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM* vm, void* reserved) {
-  JNIEnv* env = nullptr;
-  if (vm->GetEnv(reinterpret_cast<void**>(&env), JNI_VERSION) != JNI_OK) {
-    return -1;
-  }
-
-  set_vm(vm);
-  // At this point, we know Android APIs are available. Register cert chain validation JNI calls.
-  envoy_status_t result = jvm_register_cert_validator();
-  if (result == ENVOY_FAILURE) {
-    return -1;
-  }
-  return JNI_VERSION;
 }
